requirements.txt
Flask==3.0.3
google-auth==2.35.0
requests==2.32.3